+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Group changes to describe their impact on the project, as follows:
 ### Added
 - Showing short term presence for contacts whom publish it + added setting to disable it (enabled by default for sip.linphone.org accounts)
 - Confirmation dialog before removing account
+- Attended transfer instead of blind transfer if there is more than 1 call
 
 ### Changed
 - Account EXPIRES is now set to 1 month instead of 1 year for sip.linphone.org accounts